+++ geocoding/src/exports/client.ts
@@ -1,1 +1,0 @@
-export { GeocodingFieldComponent } from '../components/GeocodingFieldComponent'