+++ pages/src/exports/server.ts
@@ -1,3 +1,3 @@
 export { IsRootPageField } from '../components/server/IsRootPageField.js'
 export { ParentField } from '../components/server/ParentField.js'
-export { SlugFieldWrapper } from '../components/server/SlugFieldWrapper.js'
+export { SlugField } from '../components/server/SlugField.js'