+++ pages/src/utils/setPageVirtualFields.ts
@@ -2,11 +2,8 @@ import { PayloadRequest } from 'payload'
 import { Breadcrumb } from '../types/Breadcrumb.js'
 import { Locale } from '../types/Locale.js'
 import { PageCollectionConfigAttributes } from '../types/PageCollectionConfigAttributes.js'
-import { PagesPluginConfig } from '../types/PagesPluginConfig.js'
 import { SeoMetadata } from '../types/SeoMetadata.js'
 import { getBreadcrumbs } from './getBreadcrumbs.js'
-import { validateBreadcrumbs } from './validateBreadcrumbs.js'
-import { validatePath } from './validatePath.js'
 
 /** Sets the virtual fields (breadcrumbs, path, alternatePaths) of the given root page document. */
 export async function setPageDocumentVirtualFields({
@@ -56,12 +53,6 @@ export async function setPageDocumentVirtualFields({
     )
 
     if (locale === 'all') {
-      // TODO: remove these validations in favor of more unit tests?
-      Object.entries(paths).forEach(([_, path]) => validatePath(path, doc.id, 'all'))
-      Object.entries(paths).forEach(([locale, _]) =>
-        validateBreadcrumbs(locale as Locale, breadcrumbs[locale as Locale]),
-      )
-
       return {
         ...doc,
         path: paths,
@@ -72,10 +63,6 @@ export async function setPageDocumentVirtualFields({
         },
       }
     } else {
-      // TODO: remove these validations in favor of more unit tests?
-      validatePath(paths[locale], doc.id, locale)
-      validateBreadcrumbs(locale, breadcrumbs[locale])
-
       return {
         ...doc,
         path: paths[locale],
@@ -96,10 +83,6 @@ export async function setPageDocumentVirtualFields({
       data: doc,
       locale: undefined,
     })) as Breadcrumb[]
-
-    // TODO: remove these validations in favor of more unit tests?
-    validatePath(breadcrumbs.at(-1)!.path, doc.id, locale)
-    validateBreadcrumbs(locale, breadcrumbs)
 
     return {
       ...doc,